Move the file sharing app to migration
