+++ apps/files_sharing/appinfo/info.xml
@@ -51,4 +51,10 @@ Turning the feature off removes shared files and folders on the server for all s
 	<commands>
 		<command>OCA\Files_Sharing\Command\CleanupRemoteStorages</command>
 	</commands>
+
+	<repair-steps>
+		<post-migration>
+			<step>OCA\Files_Sharing\Migration\SetPasswordColumn</step>
+		</post-migration>
+	</repair-steps>
 </info>
