--- apps/files_sharing/appinfo/update.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * @copyright Copyright (c) 2016, ownCloud, Inc.
- *
- * @author Björn Schießle <bjoern@schiessle.org>
- * @author Roeland Jago Douma <roeland@famdouma.nl>
- *
- * @license AGPL-3.0
- *
- * This code is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License, version 3,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License, version 3,
- * along with this program.  If not, see <http://www.gnu.org/licenses/>
- *
- */
-use OCA\Files_Sharing\Migration;
-
-$installedVersion = \OC::$server->getConfig()->getAppValue('files_sharing', 'installed_version');
-
-if (version_compare($installedVersion, '1.4.0', '<')) {
-	$m = new Migration(\OC::$server->getDatabaseConnection(), \OC::$server->getConfig());
-	$m->addPasswordColumn();
-}
